Create transition for the navigation
