+++ django-website/frontend/client/js/index.js
@@ -6,7 +6,6 @@ require('../../node_modules/waypoints/lib/noframework.waypoints.js');
 
 // Import JavaScript custom files
 require('./scripts/nav.js');
-require('./scripts/expand-logo.js');
 require('./scripts/reveal-subtitles.js');
 require('./scripts/reveal-services.js');
 require('./scripts/reveal-sliding.js');
