+++ django-website/templates/partials/header.html
@@ -5,7 +5,11 @@
     <a class="header-main_interactions-logo" href="{% slugurl 'home' %}">
       {% include "partials/header-logo.html" %}
     </a>
-    <div class="button open_nav js-toggle_nav">Menu</div>
-    <div class="button close_nav js-toggle_nav">Close</div>
+    <div class="button open_nav js-toggle_nav">
+      <span class="js-toggle_nav">Menu</span>
+    </div>
+    <div class="button close_nav js-toggle_nav">
+      <span class="js-toggle_nav">Close</span>
+    </div>
   </div>
 </header>
