+++ django-website/templates/partials/nav.html
@@ -2,11 +2,11 @@
 
 <nav class="nav js-toggle_nav">
   <ul class="nav-items">
-    <li><a class="button glow_link is-line_appearing" href="{% slugurl 'agency' %}">Agency</a></li>
-    <li><a class="button glow_link is-line_appearing" href="{% slugurl 'works' %}">Works</a></li>
-    <li><a class="button glow_link is-line_appearing" href="{% slugurl 'careers' %}">Careers</a></li>
-    <li><a class="button glow_link is-line_appearing" href="{% slugurl 'contacts' %}">Contacts</a></li>
-    <li><a class="button glow_link is-line_appearing" href="{% slugurl 'blog' %}">Blog</a></li>
+    <li class="js-toggle_nav"><a class="button glow_link is-line_appearing" href="{% slugurl 'agency' %}">Agency</a></li>
+    <li class="js-toggle_nav"><a class="button glow_link is-line_appearing" href="{% slugurl 'works' %}">Works</a></li>
+    <li class="js-toggle_nav"><a class="button glow_link is-line_appearing" href="{% slugurl 'careers' %}">Careers</a></li>
+    <li class="js-toggle_nav"><a class="button glow_link is-line_appearing" href="{% slugurl 'contacts' %}">Contacts</a></li>
+    <li class="js-toggle_nav"><a class="button glow_link is-line_appearing" href="{% slugurl 'blog' %}">Blog</a></li>
   </ul>
   <ul class="nav-social">
     <li>
